// src/components/utils/form/types/DateInputTypes.ts
import { input } from "./FormTypes";

export interface dateInput extends input {
    defaultValue?: string;
    dayMin?: day;
    dayMax?: day;
    monthMin?: month;
    monthMax?: month;
    yearMin?: year;
    yearMax?: year;
}

type oneYear = 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9;
export type year = `${oneYear}${oneYear}${oneYear}${oneYear}`;
export type month = 1 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12;
export type day =
    | 1
    | 2
    | 3
    | 4
    | 5
    | 6
    | 7
    | 8
    | 9
    | 10
    | 11
    | 12
    | 13
    | 14
    | 15
    | 16
    | 17
    | 18
    | 19
    | 20
    | 21
    | 22
    | 23
    | 24
    | 25
    | 26
    | 27
    | 28
    | 29
    | 30
    | 31;